refactor(aigateway): split the provider-plugin contract into a package
core/plugin_base.py had grown to 575 lines against the 450-line limit the plan
binds to touched Python files.

Unlike the contract vocabulary module, this one cannot be brought under the limit
by moving whole definitions out: ProviderPluginBase alone is 414 lines, so
relocating every value type AND both duck-typed resolvers still leaves 456 lines
of class plus imports. The class itself has to divide.

It divides along the seam the file already marked with a section comment. The
OME-479 chat-parameter contract hooks are one port -- the sole provider-specific
input to the /v1/models summary, the detailed contract and dispatch -- and the
auth, model-registration and dispatch surface is another. ProviderPluginCore
holds the second, ProviderPluginBase extends it with the first, and plugins go on
subclassing ProviderPluginBase exactly as before.

The contract half is the SUBCLASS rather than a sibling mixin because its
derivations read the capability declarations in the other half:
available_auth_modes calls supports_api_key and oauth_config, and
chat_transport_capabilities calls supports_chat_streaming. A mixin would have to
redeclare all three, giving each of those defaults a second definition that could
drift. Inheritance needs none, and nothing in the codebase inspects __mro__,
__bases__ or __subclasses__, so the extra level is unobservable.

ProviderPluginCore is deliberately NOT re-exported: the package surface stays the
same 9 names, verified against the previous revision, and the class keeps the
same 36 members with nothing defined in both halves to shadow a default.

Behaviour-preserving: no test was added, changed or removed.

Refs: OME-653
